Correct bit shifts when interleaving multiple indexes
Previously when combining multiple classic indexes, the interleaving
process shifts input byte's bits in wrong direction. We should right
shift the input byte to append the leading bits of the input byte to
the trailing bits of the output byte. And left shift the input byte
to get the trailing bits of the input byte to the leading bits of
the next output byte.